+++ src/main/java/com/example/aereopuerto/repository/AvionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AvionRepository extends JpaRepository<Avion, Long> {
+public interface AvionRepository extends JpaRepository<Avion, Integer> {
 }
