+++ src/main/java/com/example/aereopuerto/repository/FacturaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FacturaRepository extends JpaRepository<Factura, Long> {
+public interface FacturaRepository extends JpaRepository<Factura, Integer> {
 }
